#40 Add ComponentHandler which remembers entityId
The entity id is not the same as the entity id of the world. It's only used for communication. It has to be handled clientside and serverside on its own somehow. Server world can expose its entity numbers cause it's the "parent" but client has to do it in a different way.
Not sure if it works yet.